minor, refactoring orion opening base field, need to work on orion opening simd field
