+++ poly_commit/src/orion/pcs_impl.rs
@@ -10,76 +10,7 @@ use crate::{traits::TensorCodeIOPPCS, PCS_SOUNDNESS_BITS};
 use super::utils::*;
 
 #[inline(always)]
-fn transpose_and_pack<F, PackF>(
-    evaluations: &mut [F],
-    row_num: usize,
-    msg_size: usize,
-) -> Vec<PackF>
-where
-    F: Field,
-    PackF: SimdField<Scalar = F>,
-{
-    // NOTE: pre transpose evaluations
-    let mut scratch = vec![F::ZERO; evaluations.len()];
-    transpose_in_place(evaluations, &mut scratch, row_num);
-    drop(scratch);
-
-    // NOTE: SIMD pack each row of transposed matrix
-    assert_eq!(evaluations.len() % PackF::PACK_SIZE, 0);
-    let mut packed_evals: Vec<PackF> = evaluations
-        .chunks(PackF::PACK_SIZE)
-        .map(SimdField::pack)
-        .collect();
-
-    // NOTE: transpose back to rows of evaluations, but packed
-    let packed_rows = row_num / PackF::PACK_SIZE;
-    assert_eq!(row_num % PackF::PACK_SIZE, 0);
-
-    let mut scratch = vec![PackF::ZERO; packed_rows * msg_size];
-    transpose_in_place(&mut packed_evals, &mut scratch, msg_size);
-    drop(scratch);
-
-    packed_evals
-}
-
-#[inline(always)]
-fn transpose_and_pack_from_simd<F, CircuitF, PackF>(
-    evaluations: &mut [CircuitF],
-    row_num: usize,
-    msg_size: usize,
-) -> Vec<PackF>
-where
-    F: Field,
-    CircuitF: SimdField<Scalar = F>,
-    PackF: SimdField<Scalar = F>,
-{
-    // NOTE: pre transpose evaluations
-    let mut scratch = vec![CircuitF::ZERO; evaluations.len()];
-    transpose_in_place(evaluations, &mut scratch, row_num);
-    drop(scratch);
-
-    // NOTE: SIMD pack each row of transposed matrix
-    let relative_pack_size = PackF::PACK_SIZE / CircuitF::PACK_SIZE;
-    assert_eq!(PackF::PACK_SIZE % CircuitF::PACK_SIZE, 0);
-    assert_eq!(evaluations.len() % relative_pack_size, 0);
-    let mut packed_evals: Vec<PackF> = evaluations
-        .chunks(relative_pack_size)
-        .map(SimdField::pack_from_simd)
-        .collect();
-
-    // NOTE: transpose back to rows of evaluations, but packed
-    let packed_rows = row_num / relative_pack_size;
-    assert_eq!(row_num % relative_pack_size, 0);
-
-    let mut scratch = vec![PackF::ZERO; packed_rows * msg_size];
-    transpose_in_place(&mut packed_evals, &mut scratch, msg_size);
-    drop(scratch);
-
-    packed_evals
-}
-
-#[inline(always)]
-fn commit_encoded<F, PackF>(
+pub(crate) fn commit_encoded<F, PackF>(
     pk: &OrionSRS,
     packed_evals: &[PackF],
     scratch_pad: &mut OrionScratchPad<F, PackF>,
@@ -116,121 +47,23 @@ where
     Ok(scratch_pad.interleaved_alphabet_commitment.root())
 }
 
-pub fn orion_commit_base<F, ComPackF>(
+#[inline(always)]
+pub(crate) fn orion_mt_openings<F, EvalF, ComPackF, T>(
     pk: &OrionSRS,
-    poly: &MultiLinearPoly<F>,
-    scratch_pad: &mut OrionScratchPad<F, ComPackF>,
-) -> OrionResult<OrionCommitment>
-where
-    F: Field,
-    ComPackF: SimdField<Scalar = F>,
-{
-    let (row_num, msg_size) = OrionSRS::evals_shape::<F>(poly.get_num_vars());
-    let packed_rows = row_num / ComPackF::PACK_SIZE;
-    let mut evals = poly.coeffs.clone();
-
-    let packed_evals = transpose_and_pack(&mut evals, row_num, msg_size);
-
-    commit_encoded(pk, &packed_evals, scratch_pad, packed_rows, msg_size)
-}
-
-pub fn orion_commit_simd_field<F, CircuitF, ComPackF>(
-    pk: &OrionSRS,
-    poly: &MultiLinearPoly<CircuitF>,
-    scratch_pad: &mut OrionScratchPad<F, ComPackF>,
-) -> OrionResult<OrionCommitment>
-where
-    F: Field,
-    CircuitF: SimdField<Scalar = F>,
-    ComPackF: SimdField<Scalar = F>,
-{
-    let (row_num, msg_size) = OrionSRS::evals_shape::<CircuitF>(poly.get_num_vars());
-    let relative_pack_size = ComPackF::PACK_SIZE / CircuitF::PACK_SIZE;
-    let packed_rows = row_num / relative_pack_size;
-    let mut evals = poly.coeffs.clone();
-
-    let packed_evals =
-        transpose_and_pack_from_simd::<F, CircuitF, ComPackF>(&mut evals, row_num, msg_size);
-
-    commit_encoded(pk, &packed_evals, scratch_pad, packed_rows, msg_size)
-}
-
-pub fn orion_open<F, EvalF, ComPackF, OpenPackF, T>(
-    pk: &OrionSRS,
-    poly: &MultiLinearPoly<F>,
-    point: &[EvalF],
+    leaf_range: usize,
     transcript: &mut T,
     scratch_pad: &OrionScratchPad<F, ComPackF>,
-) -> (EvalF, OrionProof<EvalF>)
+) -> Vec<tree::RangePath>
 where
     F: Field,
     EvalF: Field + From<F> + Mul<F, Output = EvalF>,
     ComPackF: SimdField<Scalar = F>,
-    OpenPackF: SimdField<Scalar = F>,
     T: Transcript<EvalF>,
 {
-    let (row_num, msg_size) = OrionSRS::evals_shape::<F>(poly.get_num_vars());
-    let num_of_vars_in_msg = msg_size.ilog2() as usize;
-
-    // NOTE: transpose evaluations for linear combinations in evaulation/proximity tests
-    let mut transposed_evaluations = poly.coeffs.clone();
-    let mut scratch = vec![F::ZERO; 1 << poly.get_num_vars()];
-    transpose_in_place(&mut transposed_evaluations, &mut scratch, row_num);
-    drop(scratch);
-
-    // NOTE: SIMD pack each row of transposed matrix
-    assert_eq!(transposed_evaluations.len() % OpenPackF::PACK_SIZE, 0);
-    let packed_evals: Vec<OpenPackF> = transposed_evaluations
-        .chunks(OpenPackF::PACK_SIZE)
-        .map(OpenPackF::pack)
-        .collect();
-    drop(transposed_evaluations);
-
-    // NOTE: declare the look up tables for column sums
-    let packed_rows = row_num / OpenPackF::PACK_SIZE;
-    let mut luts = SubsetSumLUTs::new(OpenPackF::PACK_SIZE, packed_rows);
-
-    // NOTE: working on evaluation response of tensor code IOP based PCS
-    let mut eval_row = vec![EvalF::ZERO; msg_size];
-
-    let eq_col_coeffs = EqPolynomial::build_eq_x_r(&point[num_of_vars_in_msg..]);
-    luts.build(&eq_col_coeffs);
-
-    packed_evals
-        .chunks(packed_rows)
-        .zip(eval_row.iter_mut())
-        .for_each(|(p_col, res)| *res = luts.lookup_and_sum(p_col));
-
-    // NOTE: draw random linear combination out
-    // and compose proximity response(s) of tensor code IOP based PCS
-    let proximity_test_num = pk.proximity_repetitions::<EvalF>(PCS_SOUNDNESS_BITS);
-    let mut proximity_rows = vec![vec![EvalF::ZERO; msg_size]; proximity_test_num];
-
-    proximity_rows.iter_mut().for_each(|row_buffer| {
-        let random_coeffs = transcript.generate_challenge_field_elements(row_num);
-        luts.build(&random_coeffs);
-
-        packed_evals
-            .chunks(packed_rows)
-            .zip(row_buffer.iter_mut())
-            .for_each(|(p_col, res)| *res = luts.lookup_and_sum(p_col));
-    });
-    drop(luts);
-
-    // NOTE: working on evaluation on top of evaluation response
-    let mut scratch = vec![EvalF::ZERO; msg_size];
-    let eval = MultiLinearPoly::evaluate_with_buffer(
-        &eval_row,
-        &point[..num_of_vars_in_msg],
-        &mut scratch,
-    );
-    drop(scratch);
-
     // NOTE: MT opening for point queries
-    let leaf_range = row_num / tree::leaf_adic::<F>();
     let query_num = pk.query_complexity(PCS_SOUNDNESS_BITS);
     let query_indices = transcript.generate_challenge_index_vector(query_num);
-    let query_openings = query_indices
+    query_indices
         .iter()
         .map(|qi| {
             let index = *qi % pk.codeword_len();
@@ -241,16 +74,7 @@ where
                 .interleaved_alphabet_commitment
                 .range_query(left, right)
         })
-        .collect();
-
-    (
-        eval,
-        OrionProof {
-            eval_row,
-            proximity_rows,
-            query_openings,
-        },
-    )
+        .collect()
 }
 
 pub fn orion_verify<F, EvalF, ComPackF, OpenPackF, T>(
